requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.5
django-grappelli==2.14.1
joblib==0.14.1
numpy==1.18.1
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
scikit-surprise==1.1.0
scipy==1.4.1
simpleui==3.9
six==1.14.0
sqlparse==0.3.1
ujson==2.0.3
wincertstore==0.2
